models.Object Fixed
the instance lists (objects) of subclasses were mapping to Object's lists of instances (objects). This would mean that class A (Object): pass would have the same objects list as the base class Object. We wanted A to have its own list of instances specific to A. We also wanted Object to be a culmmination of every subclasses objects list.

This now works. Object(object) simply needed to have a recursive feature that will find the current class's base classes and add the new instance to those base classes list of instances, only if that base class is of Object or has an objects list.